[FIX] Resolve remaining Pylance type errors with proper solutions
correlation.py:
- Use .values to get numpy array with clear numeric types
- Explicit float() conversion for correlation matrix values

risk_math.py:
- Convert covariance/variance to float BEFORE division and comparison
- Fixes Scalar type incompatibility with operators

earnings_checker.py:
- Replace hasattr() with isinstance(pd.Timestamp) for explicit type check
- Pylance now correctly infers to_pydatetime() availability

ai_gateway.py:
- Add type: ignore[attr-defined] for google.generativeai deprecated library
- Library has incomplete type stubs but methods exist and work
- Added explanatory comment about library deprecation

judge.py:
- Extract and validate timestamp before fromisoformat()
- isinstance(str) check ensures type safety for datetime parsing

+++ src/analysis/risk_math.py
@@ -104,10 +104,11 @@ class RiskCalculator:
             if len(aligned) < 20:
                 return 1.0
 
-            covariance = aligned.iloc[:, 0].cov(aligned.iloc[:, 1])
-            variance = aligned.iloc[:, 1].var()
+            # Converter para float antes de operações - .cov() e .var() retornam Scalar
+            covariance = float(aligned.iloc[:, 0].cov(aligned.iloc[:, 1]))
+            variance = float(aligned.iloc[:, 1].var())
 
-            return float(covariance / variance) if variance > 0 else 1.0
+            return covariance / variance if variance > 0 else 1.0
 
         except Exception:
             return 1.0
